app.py
import sys
import glob
import os
import json
import re
import pandas as pd
from dotenv import load_dotenv

load_dotenv()

def get_column_names(schemas, ds_name, sorting_key="column_position"):
    column_details = sorted(schemas[ds_name], key=lambda col: col[sorting_key])
    column_names = [col["column_name"] for col in column_details]
    return column_names


def read_csv(file_path, schemas):
    file_path_list = re.split(r'[/\\]', file_path)
    ds_name = file_path_list[-2]
    column_names = get_column_names(schemas, ds_name)
    df = pd.read_csv(file_path, 
            names=column_names, 
            header=None
    )
    return df


def to_json(df, tgt_base_dir, ds_name, file_name):
    json_file_path = f"{tgt_base_dir}/{ds_name}/{file_name}"
    os.makedirs(f"{tgt_base_dir}/{ds_name}", exist_ok=True)
    df.to_json(
        json_file_path, 
        orient="records", 
        lines=True
)


def file_converter(src_base_dir, tgt_base_dir, ds_name):
    schemas = json.load(open(f"{src_base_dir}/schemas.json"))
    files = glob.glob(f"{src_base_dir}/{ds_name}/part-*")

    if len(files) == 0:
        raise NameError(f"No files found for {ds_name}")

    for file in files:
        df = read_csv(file, schemas)
        file_name = re.split(r"[/\\]", file)[-1]
        to_json(df, tgt_base_dir, ds_name, file_name)


def process_files(ds_names=None):
    src_base_dir = os.environ.get("SRC_BASE_DIR")
    tgt_base_dir = os.environ.get("TGT_BASE_DIR")
    schemas = json.load(open(f"{src_base_dir}/schemas.json"))

    if not ds_names:
        ds_names = schemas.keys()
    for ds_name in ds_names:
        try:
            print(f"Processing {ds_name}")
            file_converter(src_base_dir, tgt_base_dir, ds_name)
        except NameError as ne:
            print(ne)
            print(f"Error Processing {ds_name}")
            pass


if __name__ == "__main__":

    if len(sys.argv) == 2:
        ds_names = json.loads(sys.argv[1])
        process_files(ds_names)
    else:
        process_files() 