requirements.txt
pandas>=2.1
python-dotenv>=1.0